+++ multiplayer_game_2d/tile_map.h
@@ -63,7 +63,7 @@ struct tile_map
       int debug_lines_vertex_count = (tile_count + 1) * 8; // 4 lines per tile and 2 vertices per line so 8 vertices per tile, the additional 1 is for background tile
       sf::VertexArray debug_line_vertices(sf::Lines, debug_lines_vertex_count);
 
-      for(int i=0,tile_map_vertex=0; i < debug_lines_vertex_count; i+=8,tile_map_vertex+=4)
+      for(size_t i=0,tile_map_vertex=0; i < debug_lines_vertex_count; i+=8,tile_map_vertex+=4)
       {
         debug_line_vertices[i].position   = this->vertex_buffer[tile_map_vertex].position;
         debug_line_vertices[i+1].position = this->vertex_buffer[tile_map_vertex+1].position;
